feat: adding firebaseConfig file
